refactor: move queries outside of commands
