+++ tdp/cli/commands/service_versions.py
@@ -4,7 +4,7 @@
 import click
 from tabulate import tabulate
 
-from tdp.cli.commands.queries import get_latest_success_service_component_version_query
+from tdp.cli.queries import get_latest_success_service_component_version_query
 from tdp.cli.session import get_session_class
 from tdp.cli.utils import database_dsn
 from tdp.core.models import ServiceComponentLog
