+++ tdp/cli/queries.py
